Make example project consume pod via Podfile.
