+++ Algorithm Class/Luhn.m
@@ -8,38 +8,124 @@
 
 #import "Luhn.h"
 
+@implementation NSString (Luhn)
+
+- (BOOL) isValidCreditCardNumber {
+    return [Luhn validateString:self];
+}
+
+- (OLCreditCardType) creditCardType {
+    return [Luhn typeFromString:self];
+}
+
+@end
+
+@interface NSString (Luhn_Private)
+
+- (NSString *) formattedStringForProcessing;
+
+@end
+
+@implementation NSString (Luhn_Private)
+
+- (NSString *) formattedStringForProcessing {
+    NSCharacterSet *illegalCharacters = [[NSCharacterSet decimalDigitCharacterSet] invertedSet];
+    NSArray *components = [self componentsSeparatedByCharactersInSet:illegalCharacters];
+    return [components componentsJoinedByString:@""];
+}
+
+@end
+
 @implementation Luhn
 
-+ (BOOL) validateString:(NSString *) _string {
++ (OLCreditCardType) typeFromString:(NSString *) string {
+    BOOL valid = [string isValidCreditCardNumber];
+    if (!valid) {
+        return OLCreditCardTypeInvalid;
+    }
     
-    NSMutableString *reversedString = [NSMutableString stringWithCapacity:[_string length]];
+    NSString *formattedString = [string formattedStringForProcessing];
+    if (formattedString == nil || formattedString.length < 9) {
+        return OLCreditCardTypeInvalid;
+    }
     
-    [_string enumerateSubstringsInRange:NSMakeRange(0, [_string length]) options:(NSStringEnumerationReverse |NSStringEnumerationByComposedCharacterSequences) usingBlock:^(NSString *substring, NSRange substringRange, NSRange enclosingRange, BOOL *stop) {
+    NSArray *enums = @[@(OLCreditCardTypeAmex), @(OLCreditCardTypeDinersClub), @(OLCreditCardTypeDiscover), @(OLCreditCardTypeJCB), @(OLCreditCardTypeMastercard), @(OLCreditCardTypeVisa)];
+    
+    __block OLCreditCardType type = OLCreditCardTypeInvalid;
+    [enums enumerateObjectsUsingBlock:^(id obj, NSUInteger idx, BOOL *stop) {
+        OLCreditCardType _type = [obj integerValue];
+        NSPredicate *predicate = [Luhn predicateForType:_type];
+        BOOL isCurrentType = [predicate evaluateWithObject:formattedString];
+        if (isCurrentType) {
+            type = _type;
+            *stop = YES;
+        }
+    }];
+    return type;
+}
+
++ (NSPredicate *) predicateForType:(OLCreditCardType) type {
+    if (type == OLCreditCardTypeInvalid || type == OLCreditCardTypeUnsupported) {
+        return nil;
+    }
+    NSString *regex = nil;
+    switch (type) {
+        case OLCreditCardTypeAmex:
+            regex = @"^3[47][0-9]{5,}$";
+            break;
+        case OLCreditCardTypeDinersClub:
+            regex = @"^3(?:0[0-5]|[68][0-9])[0-9]{4,}$";
+            break;
+        case OLCreditCardTypeDiscover:
+            regex = @"^6(?:011|5[0-9]{2})[0-9]{3,}$";
+            break;
+        case OLCreditCardTypeJCB:
+            regex = @"^(?:2131|1800|35[0-9]{3})[0-9]{3,}$";
+            break;
+        case OLCreditCardTypeMastercard:
+            regex = @"^5[1-5][0-9]{5,}$";
+            break;
+        case OLCreditCardTypeVisa:
+            regex = @"^4[0-9]{6,}$";
+            break;
+        default:
+            break;
+    }
+    return [NSPredicate predicateWithFormat:@"SELF MATCHES %@", regex];
+}
+
++ (BOOL) validateString:(NSString *)string forType:(OLCreditCardType)type {
+    return [Luhn typeFromString:string] == type;
+}
+
++ (BOOL)validateString:(NSString *)string {
+    NSParameterAssert(string != nil);
+    NSParameterAssert(string.length >= 9);
+    
+    NSString *formattedString = [string formattedStringForProcessing];
+    if (formattedString == nil || formattedString.length < 9) {
+        return NO;
+    }
+    
+    NSMutableString *reversedString = [NSMutableString stringWithCapacity:[formattedString length]];
+    
+    [formattedString enumerateSubstringsInRange:NSMakeRange(0, [formattedString length]) options:(NSStringEnumerationReverse |NSStringEnumerationByComposedCharacterSequences) usingBlock:^(NSString *substring, NSRange substringRange, NSRange enclosingRange, BOOL *stop) {
         [reversedString appendString:substring];
     }];
-        
-    unsigned int odd_sum = 0, even_sum = 0;
     
-    for (int i = 0; i < [reversedString length]; i++) {
-        
+    NSUInteger oddSum = 0, evenSum = 0;
+    
+    for (NSUInteger i = 0; i < [reversedString length]; i++) {
         NSInteger digit = [[NSString stringWithFormat:@"%C", [reversedString characterAtIndex:i]] integerValue];
         
         if (i % 2 == 0) {
-            
-            odd_sum += digit;
-            
+            evenSum += digit;
         }
-        
         else {
-            
-            even_sum += digit / 5 + ( 2 * digit) % 10;
-            
+            oddSum += digit / 5 + (2 * digit) % 10;
         }
-                
     }
-        
-    return (odd_sum + even_sum) % 10 == 0;
-    
+    return (oddSum + evenSum) % 10 == 0;
 }
 
 @end
