+++ Example Project/Podfile
@@ -1,6 +1,10 @@
 platform :ios, '5.0'
 inhibit_all_warnings!
 
+target 'Luhn Algorithm (Mod 10)', :exclusive => true do
+  pod 'ObjectiveLuhn', :path => '../ObjectiveLuhn.podspec'
+end
+
 target :Tests, :exclusive => true do
-	pod 'Kiwi/XCTest'
+  pod 'Kiwi/XCTest'
 end
